experimentos/pyproject.toml
[project]
name = "experimentos"
version = "0.1.0"
description = "Hybrid RAG system for master's thesis research"
requires-python = ">=3.13"

dependencies = [
    "langgraph>=1.0.3",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-google-genai>=2.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "chromadb>=1.3.5",
    "polars>=1.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.0",
    "sqlglot>=26.0.0",
    "docling>=2.63.0",
    "typer>=0.15.0",
    "rich>=14.0.0",
    "python-dotenv>=1.2.1",
    "loguru>=0.7.3",
    "numpy>=2.0.0",
    "pyarrow>=23.0.1",
    "scikit-learn>=1.8.0",
    "pingouin>=0.6.1",
]

[project.scripts]
rag = "src.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
